refactor(pipeline): split PiiRedactor/Facade.ts into Routing + Dispatch + Composer (Phase 8.5c / C1)
Drains the §13A `max-lines: 'off'` grandfather entry for
`PiiRedactor/Facade.ts`. The 271-LoC god-facade decomposes cleanly
along its three already-distinct responsibilities — path-tail
classification, strategy dispatch, and the unified value-only entry
point. The grandfather override is left in `eslint.config.mjs` for
this commit; C2 in the same ladder removes the §13A block.

## Files

* **`src/Scrapers/Pipeline/Types/PiiRedactor/Routing.ts`** (NEW, 113 LoC)
  Owns the `PATH_TAIL_TO_CATEGORY` routing table, the case-insensitive
  `TOKEN_SUFFIXES` / `NAME_SUFFIXES` lists, the `isTokenSuffix` /
  `isNameSuffix` predicates, and the `classifyKey()` resolver. Imports
  only `PiiCategory` / `PiiClassifierBool` from `Types.js`. Zero
  strategy knowledge.

* **`src/Scrapers/Pipeline/Types/PiiRedactor/Dispatch.ts`** (NEW, 116 LoC)
  Owns the per-category `STRING_STRATEGIES` lookup table, the
  `toStringValue` / `toAmountValue` coercion helpers, the
  `dispatchStrategy` + `censorTail` pipeline (with strategy-throw
  → `REDACTION_ERROR_HINT` translation), and the Pino
  `createCensorFn()` factory. Owns the `CensorValue` / `CensorFn`
  types. Imports `classifyKey` from `Routing.js`. Zero routing-table
  knowledge.

* **`src/Scrapers/Pipeline/Types/PiiRedactor/Facade.ts`** (MUTATED, 271 → 61 LoC)
  Now the thin composer — owns the value-only `redact()` entry point
  + the inner `redactStringValue` auth-credential sniffer chain (token
  / OTP / cookie). Re-exports `classifyKey` and
  `PATH_TAIL_TO_CATEGORY` from `Routing.js`, and `CensorFn` /
  `CensorValue` / `createCensorFn` from `Dispatch.js`, so the parent
  barrel (`Types/PiiRedactor.ts`) and every external caller observe
  the SAME public surface — zero downstream churn.

## Behaviour

100% preserved. Each extracted function is moved verbatim. The
`STRING_STRATEGIES` registry, the `redact()` default-deny rule, the
auth-credential ordering inside `redactStringValue`, the
`REDACTED_HINT` / `OTP_HINT` / `REDACTION_ERROR_HINT` return values,
the empty-path / empty-tail short-circuits in `createCensorFn` — all
identical.

## Why this shape

* `Routing.ts` and `Dispatch.ts` are the two functional concerns the
  god-facade jammed together: routing maps INPUT (path-tail key) to
  CATEGORY; dispatch maps CATEGORY + VALUE to OUTPUT (stable hint).
  Their dependency direction is one-way (`Dispatch` → `Routing`); no
  cycle.
* The composer is intentionally NOT renamed away from `Facade.ts` —
  the existing barrel re-export (`from './PiiRedactor/Facade.js'`)
  and all in-tree JSDoc cross-references continue to resolve.
* The split intentionally does NOT move `Facade.ts` up to sit beside
  the parent barrel — the per-concern modules stay co-located inside
  the `PiiRedactor/` folder so a future contributor sees the full
  cluster (Account, Amount, AuthCredentials, Card, Cookie, …,
  Routing, Dispatch, Facade) in one directory listing.

## Validation

* `npm run type-check` — passes.
* `npx eslint src --max-warnings 0` — clean (all 49 TS + 5 bash
  canaries still tripping their intended rules).
* `node node_modules/jest/bin/jest.js --testPathPatterns=PiiRedactor` — 145/145 passing.
  The two cluster invariants (`DefaultDeny.test.ts`,
  `CredentialsNoBypass.test.ts`) continue to call `redact()` through
  the parent barrel — no test churn.

## Plan tracking

* Master: `pipeline-decoupling-master-2026-05-28/status.txt`
* Sub-phase: `phase-8.5-canonical-10/sub-c-pii-types-docs/status.txt`
* Ladder position: C1 / 10. Drains the §13A `Facade.ts` grandfather
  entry. C2 (next commit) removes the §13A override block from
  `eslint.config.mjs` AND adds `max-lines-per-function:10` to §7b/§12e
  to enforce the global ≤10-LoC function cap going forward.
* Prior commits in this ladder: T1 `67bd166b` (canary infra), T2
  `b614b638` (SeedCarryBootstrapKind unify), T3 `4c92e2c7`
  (IProbeConfig strict XOR), T4 `584fa230` (JsonValueTemplate
  recursive + array arm).
